+++ tests/recursive_logging.rs
@@ -17,8 +17,6 @@ use logforth::append::rolling_file::NonBlockingBuilder;
 use logforth::append::rolling_file::RollingFileWriter;
 use logforth::append::rolling_file::Rotation;
 use logforth::layout;
-use logforth::Dispatch;
-use logforth::Logger;
 
 // ensure logforth's impl doesn't properly handle recursive logging
 #[test]
@@ -39,13 +37,13 @@ fn test_meta_logging_in_format_works() {
         })
     };
 
-    Logger::new()
-        .dispatch(Dispatch::new().append(append::Stdout::default().with_layout(layout("out"))))
-        .dispatch(Dispatch::new().append(append::Stderr::default().with_layout(layout("err"))))
-        .dispatch(
-            Dispatch::new().append(append::RollingFile::new(writer).with_layout(layout("file"))),
-        )
-        .apply()
+    logforth::builder()
+        .append(append::Stdout::default().with_layout(layout("out")))
+        .dispatch()
+        .append(append::Stderr::default().with_layout(layout("err")))
+        .dispatch()
+        .append(append::RollingFile::new(writer).with_layout(layout("file")))
+        .finish()
         .unwrap();
 
     struct Thing<'a>(&'a str);
